Virus/Update/Virus/requirements.txt
pyinstaller==5.1
cryptography==38.0.4
pycryptodome==3.16.0
pywin32==305
requests==2.28.1
urllib3==1.26.13